Removes accidentally added directory src
