--- demo/environments/environment.dev.ts
@@ -1,3 +1,0 @@
-export const environment = {
-  production: false
-};
