+++ demo/polyfills.ts
@@ -1,4 +1,5 @@
-// Prefer CoreJS over the polyfills above
+// This file includes polyfills needed by Angular and is loaded before
+// the app. You can add your own extra polyfills to this file.
 import 'core-js/es6/symbol';
 import 'core-js/es6/object';
 import 'core-js/es6/function';
