+++ demo/tsconfig.json
@@ -1,21 +1,18 @@
 {
-    "compilerOptions": {
-        "declaration": false,
-        "emitDecoratorMetadata": true,
-        "experimentalDecorators": true,
-        "lib": ["es6", "dom"],
-        "mapRoot": "./",
-        "module": "es6",
-        "moduleResolution": "node",
-        "outDir": "../dist/out-tsc",
-        "sourceMap": true,
-        "target": "es5",
-        "typeRoots": [
-            "../node_modules/@types"
-        ],
-        "types": [
-            "requirejs",
-            "prismjs"
-        ]
-    }
+  "compilerOptions": {
+    "baseUrl": "",
+    "declaration": false,
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
+    "lib": ["es6", "dom"],
+    "mapRoot": "./",
+    "module": "es6",
+    "moduleResolution": "node",
+    "outDir": "../dist/out-tsc",
+    "sourceMap": true,
+    "target": "es5",
+    "typeRoots": [
+      "../node_modules/@types"
+    ]
+  }
 }
